requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.17.2
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
torch==1.2.0
torchvision==0.4.0